backend/app/api/schemas/user.py
from pydantic import BaseModel, EmailStr, Field
from datetime import datetime

class RegisterRequest(BaseModel):
    email: EmailStr
    name: str = Field(min_length=1, max_length=255)
    password: str = Field(min_length=2)


class LoginRequest(BaseModel):
    email: EmailStr
    password: str


class TokenResponse(BaseModel):
    access_token:str
    token_type: str = "bearer"


class UserResponse(BaseModel):
    id: int 
    email: EmailStr
    name: str
    created_at: datetime 

    class Config:
        from_attributes = True